test(Tab): update test cases
